requirements.txt
streamlit==1.41.1
openai-whisper==20240930
pydub==0.25.1


